--- app/store/index.ts
@@ -1,2 +1,0 @@
-export * from '../lib/compute-score'
-export * from '~/consts'
